requirements.txt
# GUI
PyQt5>=5.15.0

# 核心依赖
numpy>=1.19.0
open3d>=0.13.0
pillow>=8.0.0

# 深度学习
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.12  # MiDaS依赖

# 图像处理
opencv-python>=4.5.0
scipy>=1.7.0

# 工具和实用程序
tqdm>=4.65.0  # 进度条
matplotlib>=3.3.0  # 可视化
scikit-learn>=0.24.0  # 点云处理

# 可选依赖（推荐安装）
cuda-python  # GPU加速支持
ninja  # 编译加速
